refactored questionaire questions
- ensured that ideas questions are correct
