--- src/QualitySliderQuestions.js
@@ -1,1131 +1,0 @@
-export const QualitySlider = {
-	development: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "What kind of code contribution is this?",
-				answers: [
-				{
-					answer: "Serius Security fix",
-					value: 30,
-				},
-				{
-					answer: "Completely new project",
-					value: 30,
-				},
-				{
-					answer: "Contains Bugfixes and new features",
-					value: 30,
-				},
-				{
-					answer: "New Feature",
-					value: 25,
-				},
-				{
-					answer: "Other",
-					value: 15,
-				}],
-			},
-			{
-				question: "How important is the feature / bugfix for the project?",
-				answers: [
-				{
-					answer: "Very important",
-					value: 30,
-				},
-				{
-					answer: "Important",
-					value: 20,
-				},
-				{
-					answer: "Less Important",
-					value: 15,
-				},
-				{
-					answer: "Not important",
-					value: 10,
-				}],
-			}
-		]
-	},
-	blog: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is there visual content in the contribution?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 25,
-				},
-				{
-					answer: "Yes, but bad quality",
-					value: 10,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is this a unique blog post?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 25,
-				},
-				{
-					answer: "Not so unique",
-					value: 10,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is the blog post well structured?",
-				answers: [
-				{
-					answer: "Yes, very nice",
-					value: 10,
-				},
-				{
-					answer: "Not so good",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-		]
-	},
-	tutorials: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is there information about the additional frameworks?",
-				answers: [
-				{
-					answer: "Yes, everything is explained",
-					value: 10,
-				},
-				{
-					answer: "Yes, but not enough",
-					value: 5,
-				},
-				{
-					answer: "No details at all",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is there code in the tutorial?",
-				answers: [
-				{
-					answer: "Yes, and it’s well explained",
-					value: 20,
-				},
-				{
-					answer: "Yes, but no explanation",
-					value: 10,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is there visual content in the contribution?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Yes, but bad quality",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is this a tutorial series?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Yes, but first part",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is the tutorial post structured?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Not so good",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-		]
-	},
-	"video-tutorials": {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Length of video?",
-				answers: [
-				{
-					answer: "More than 15 minutes",
-					value: 60,
-				},
-				{
-					answer: "More than 10 minutes",
-					value: 30,
-				},
-				{
-					answer: "Around 5 minutes",
-					value: 10,
-				}],
-			},
-		]
-	},
-	translations: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "How many translated words?",
-				answers: [
-				{
-					answer: "2000+",
-					value: 10,
-				},
-				{
-					answer: "1500-2000",
-					value: 5,
-				},
-				{
-					answer: "1000-1500",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Are there any traces of machine translation making the work inaccurate?",
-				answers: [
-				{
-					answer: "No",
-					value: 20,
-				},
-				{
-					answer: "Yes",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is there code that shouldn’t be translated?",
-				answers: [
-				{
-					answer: "No",
-					value: 25,
-				},
-				{
-					answer: "Yes",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is this a proofreading contribution?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			}
-		]
-	},
-	graphics: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "This contribution is a ...",
-				answers: [
-				{
-					answer: "Marged / Used art work",
-					value: 10,
-				},
-				{
-					answer: "Proposal",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "How many art work is available in the contribution?",
-				answers: [
-				{
-					answer: "10+",
-					value: 10,
-				},
-				{
-					answer: "5-10",
-					value: 7,
-				},
-				{
-					answer: "2-3",
-					value: 5,
-				},
-				{
-					answer: "1",
-					value: 2,
-				},
-				],
-			},
-			{
-				question: "Is the license present for the contribution?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Quality of images?",
-				answers: [
-				{
-					answer: "Good Quality",
-					value: 20,
-				},
-				{
-					answer: "Acceptable",
-					value: 10,
-				},
-				{
-					answer: "Bad Quality",
-					value: 0,
-				}
-				],
-			}
-		]
-	},
-	"bug-hunting": {
-		questions: [
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the bug report formal / informal?",
-				answers: [
-				{
-					answer: "Yes straight to the point",
-					value: 50,
-				},
-				{
-					answer: "No steps to reproduce",
-					value: 25,
-				},
-				{
-					answer: "Not informal and not formal",
-					value: 0,
-				}
-				],
-			},
-			{
-				question: "Is there visual presentation provided for the bug?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Yes but low quality",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				},
-				],
-			},
-			{
-				question: "Is there any unrelated content in the bug report?",
-				answers: [
-				{
-					answer: "No, post solely discusses only talk about the bug report",
-					value: 10,
-				},
-				{
-					answer: "Yes, personal intro or other unrelated content ",
-					value: 0,
-				}
-				],
-			}
-		]
-	},
-	social: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Kind of visibility?",
-				answers: [
-				{
-					answer: "Paid Advertisement",
-					value: 60,
-				},
-				{
-					answer: "Meetup, Group Gathering",
-					value: 50,
-				},
-				{
-					answer: "Presentation, or Youtube",
-					value: 30,
-				},
-				{
-					answer: "Other",
-					value: 20,
-				}
-				],
-			}
-		]
-	},
-	documentation: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the documentation good in detail?",
-				answers: [
-				{
-					answer: "Yes, lot's of details and examples",
-					value: 60,
-				},
-				{
-					answer: "Yes lot's of details but not many examples",
-					value: 30,
-				},
-				{
-					answer: "Not much details",
-					value: 0,
-				}
-				],
-			},
-		]
-	},
-	analysis: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the analysis benefical in the future?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 60,
-				},
-				{
-					answer: "Partially",
-					value: 30,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-		]
-	},
-	ideas: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-		]
-	},
-	"sub-projects": {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-		]
-	},
-	copywriting: {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-		]
-	},
-	'task-development': {
-		questions: [
-			{
-				question: "Is the project description formal?",
-				answers: [
-				{
-					answer: "Yes it’s straight to the point",
-					value: 10,
-				},
-				{
-					answer: "Need more description ",
-					value: 5,
-				},
-				{
-					answer: "Not too descriptive",
-					value: 0,
-				}],
-			},
-			{
-				question: "Is the language / grammar correct?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 20,
-				},
-				{
-					answer: "A few mistakes",
-					value: 10,
-				},
-				{
-					answer: "It's pretty bad",
-					value: 0,
-				}],
-			},
-			{
-				question: "Was the template followed?",
-				answers: [
-				{
-					answer: "Yes",
-					value: 10,
-				},
-				{
-					answer: "Partially",
-					value: 5,
-				},
-				{
-					answer: "No",
-					value: 0,
-				}],
-			},
-		]
-	},
-};
