+++ src/components/Story/questionaire/blog.js
@@ -1,0 +1,103 @@
+export default {
+	questions: [
+		{
+			question: "Is the project description formal?",
+			answers: [
+			{
+				answer: "Yes it’s straight to the point",
+				value: 10,
+			},
+			{
+				answer: "Need more description ",
+				value: 5,
+			},
+			{
+				answer: "Not too descriptive",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the language / grammar correct?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 20,
+			},
+			{
+				answer: "A few mistakes",
+				value: 10,
+			},
+			{
+				answer: "It's pretty bad",
+				value: 0,
+			}],
+		},
+		{
+			question: "Was the template followed?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is there visual content in the contribution?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 25,
+			},
+			{
+				answer: "Yes, but bad quality",
+				value: 10,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is this a unique blog post?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 25,
+			},
+			{
+				answer: "Not so unique",
+				value: 10,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is the blog post well structured?",
+			answers: [
+			{
+				answer: "Yes, very nice",
+				value: 10,
+			},
+			{
+				answer: "Not so good",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+	]
+};
