+++ src/components/Story/questionaire/development.js
@@ -1,0 +1,96 @@
+export default {
+	questions: [
+		{
+			question: "Is the project description formal?",
+			answers: [
+			{
+				answer: "Yes it’s straight to the point",
+				value: 10,
+			},
+			{
+				answer: "Need more description ",
+				value: 5,
+			},
+			{
+				answer: "Not too descriptive",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the language / grammar correct?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 20,
+			},
+			{
+				answer: "A few mistakes",
+				value: 10,
+			},
+			{
+				answer: "It's pretty bad",
+				value: 0,
+			}],
+		},
+		{
+			question: "Was the template followed?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "What kind of code contribution is this?",
+			answers: [
+			{
+				answer: "Serius Security fix",
+				value: 30,
+			},
+			{
+				answer: "Completely new project",
+				value: 30,
+			},
+			{
+				answer: "Contains Bugfixes and new features",
+				value: 30,
+			},
+			{
+				answer: "New Feature",
+				value: 25,
+			},
+			{
+				answer: "Other",
+				value: 15,
+			}],
+		},
+		{
+			question: "How important is the feature / bugfix for the project?",
+			answers: [
+			{
+				answer: "Very important",
+				value: 30,
+			},
+			{
+				answer: "Important",
+				value: 20,
+			},
+			{
+				answer: "Less Important",
+				value: 15,
+			},
+			{
+				answer: "Not important",
+				value: 10,
+			}],
+		}
+	]
+};
