+++ src/components/Story/questionaire/graphics.js
@@ -1,0 +1,116 @@
+export default {
+    questions: [
+        {
+            question: "Is the project description formal?",
+            answers: [
+            {
+                answer: "Yes it’s straight to the point",
+                value: 10,
+            },
+            {
+                answer: "Need more description ",
+                value: 5,
+            },
+            {
+                answer: "Not too descriptive",
+                value: 0,
+            }],
+        },
+        {
+            question: "Is the language / grammar correct?",
+            answers: [
+            {
+                answer: "Yes",
+                value: 20,
+            },
+            {
+                answer: "A few mistakes",
+                value: 10,
+            },
+            {
+                answer: "It's pretty bad",
+                value: 0,
+            }],
+        },
+        {
+            question: "Was the template followed?",
+            answers: [
+            {
+                answer: "Yes",
+                value: 10,
+            },
+            {
+                answer: "Partially",
+                value: 5,
+            },
+            {
+                answer: "No",
+                value: 0,
+            }],
+        },
+        {
+            question: "This contribution is a ...",
+            answers: [
+            {
+                answer: "Marged / Used art work",
+                value: 10,
+            },
+            {
+                answer: "Proposal",
+                value: 0,
+            }
+            ],
+        },
+        {
+            question: "How many art work is available in the contribution?",
+            answers: [
+            {
+                answer: "10+",
+                value: 10,
+            },
+            {
+                answer: "5-10",
+                value: 7,
+            },
+            {
+                answer: "2-3",
+                value: 5,
+            },
+            {
+                answer: "1",
+                value: 2,
+            },
+            ],
+        },
+        {
+            question: "Is the license present for the contribution?",
+            answers: [
+            {
+                answer: "Yes",
+                value: 20,
+            },
+            {
+                answer: "No",
+                value: 0,
+            }
+            ],
+        },
+        {
+            question: "Quality of images?",
+            answers: [
+            {
+                answer: "Good Quality",
+                value: 20,
+            },
+            {
+                answer: "Acceptable",
+                value: 10,
+            },
+            {
+                answer: "Bad Quality",
+                value: 0,
+            }
+            ],
+        }
+    ]
+};
