+++ src/components/Story/questionaire/ideas.js
@@ -1,0 +1,112 @@
+export default {
+	questions: [
+		{
+			question: "Is the project description formal?",
+			answers: [
+			{
+				answer: "Yes it’s straight to the point",
+				value: 10,
+			},
+			{
+				answer: "Need more description ",
+				value: 5,
+			},
+			{
+				answer: "Not too descriptive",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the language / grammar correct?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 20,
+			},
+			{
+				answer: "A few mistakes",
+				value: 10,
+			},
+			{
+				answer: "It's pretty bad",
+				value: 0,
+			}],
+		},
+		{
+			question: "Was the template followed?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the suggestion related to any technical issue?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the suggestion descriptive enough?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is there any unrelated content in the suggestion?",
+			answers: [
+			{
+				answer: "No, post solely discusses only talk about the suggestion",
+				value: 10,
+			},
+			{
+				answer: "Yes, personal intro or other unrelated content",
+				value: 0,
+			}],
+		},
+		{
+			question: "How would you rate this suggestion as in the impact it could have on the project?",
+			answers: [
+			{
+				answer: "Excellent",
+				value: 15,
+			},
+			{
+				answer: "Very good",
+				value: 12,
+			},
+			{
+				answer: "Good",
+				value: 8,
+			},
+			{
+				answer: "Basic enhancement",
+				value: 2,
+			}],
+		},
+	]
+};
