+++ src/components/Story/questionaire/social.js
@@ -1,0 +1,73 @@
+export default {
+	questions: [
+		{
+			question: "Is the project description formal?",
+			answers: [
+			{
+				answer: "Yes it’s straight to the point",
+				value: 10,
+			},
+			{
+				answer: "Need more description ",
+				value: 5,
+			},
+			{
+				answer: "Not too descriptive",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the language / grammar correct?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 20,
+			},
+			{
+				answer: "A few mistakes",
+				value: 10,
+			},
+			{
+				answer: "It's pretty bad",
+				value: 0,
+			}],
+		},
+		{
+			question: "Was the template followed?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Kind of visibility?",
+			answers: [
+			{
+				answer: "Paid Advertisement",
+				value: 60,
+			},
+			{
+				answer: "Meetup, Group Gathering",
+				value: 50,
+			},
+			{
+				answer: "Presentation, or Youtube",
+				value: 30,
+			},
+			{
+				answer: "Other",
+				value: 20,
+			}
+			],
+		}
+	]
+};
