+++ src/components/Story/questionaire/tutorials.js
@@ -1,0 +1,137 @@
+export default {
+	questions: [
+		{
+			question: "Is the project description formal?",
+			answers: [
+			{
+				answer: "Yes it’s straight to the point",
+				value: 10,
+			},
+			{
+				answer: "Need more description ",
+				value: 5,
+			},
+			{
+				answer: "Not too descriptive",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is the language / grammar correct?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 20,
+			},
+			{
+				answer: "A few mistakes",
+				value: 10,
+			},
+			{
+				answer: "It's pretty bad",
+				value: 0,
+			}],
+		},
+		{
+			question: "Was the template followed?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Partially",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}],
+		},
+		{
+			question: "Is there information about the additional frameworks?",
+			answers: [
+			{
+				answer: "Yes, everything is explained",
+				value: 10,
+			},
+			{
+				answer: "Yes, but not enough",
+				value: 5,
+			},
+			{
+				answer: "No details at all",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is there code in the tutorial?",
+			answers: [
+			{
+				answer: "Yes, and it’s well explained",
+				value: 20,
+			},
+			{
+				answer: "Yes, but no explanation",
+				value: 10,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is there visual content in the contribution?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Yes, but bad quality",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is this a tutorial series?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Yes, but first part",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+		{
+			question: "Is the tutorial post structured?",
+			answers: [
+			{
+				answer: "Yes",
+				value: 10,
+			},
+			{
+				answer: "Not so good",
+				value: 5,
+			},
+			{
+				answer: "No",
+				value: 0,
+			}
+			],
+		},
+	]
+};
