+++ src/components/Story/questionaire/video-tutorials.js
@@ -1,0 +1,68 @@
+export default {
+    questions: [
+        {
+            question: "Is the project description formal?",
+            answers: [
+            {
+                answer: "Yes it’s straight to the point",
+                value: 10,
+            },
+            {
+                answer: "Need more description ",
+                value: 5,
+            },
+            {
+                answer: "Not too descriptive",
+                value: 0,
+            }],
+        },
+        {
+            question: "Is the language / grammar correct?",
+            answers: [
+            {
+                answer: "Yes",
+                value: 20,
+            },
+            {
+                answer: "A few mistakes",
+                value: 10,
+            },
+            {
+                answer: "It's pretty bad",
+                value: 0,
+            }],
+        },
+        {
+            question: "Was the template followed?",
+            answers: [
+            {
+                answer: "Yes",
+                value: 10,
+            },
+            {
+                answer: "Partially",
+                value: 5,
+            },
+            {
+                answer: "No",
+                value: 0,
+            }],
+        },
+        {
+            question: "Length of video?",
+            answers: [
+            {
+                answer: "More than 15 minutes",
+                value: 60,
+            },
+            {
+                answer: "More than 10 minutes",
+                value: 30,
+            },
+            {
+                answer: "Around 5 minutes",
+                value: 10,
+            }],
+        },
+    ]
+};
